+++ NEWS.md
@@ -1,6 +1,7 @@
 # phonfieldwork 0.0.12
 
 - add a `separate_duration` argument to the `concatenate_soundfiles()` function that makes it possible to use some silent separator during the file concatenation.
+- make `rename_soundfiles()` function to work with mp3 files.
 
 # phonfieldwork 0.0.11